update to golang.org/x/tools 09fb680d

+++ fastwalk/fastwalk_dirent_ino.go
@@ -2,13 +2,14 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-//go:build linux && !appengine
-// +build linux,!appengine
+//go:build (linux || (darwin && !cgo)) && !appengine
+// +build linux darwin,!cgo
+// +build !appengine
 
 package fastwalk
 
 import "syscall"
 
 func direntInode(dirent *syscall.Dirent) uint64 {
-	return uint64(dirent.Ino)
+	return dirent.Ino
 }
